requirements.txt
torch
transformers
datasets
flask
pyyaml